MAINT: merge _validate_bounds from optimize and qmc to _util
