+++ scipy/_lib/meson.build
@@ -98,6 +98,7 @@ py3.extension_module('messagestream',
 
 python_sources = [
   '__init__.py',
+  '_bounds.py',
   '_bunch.py',
   '_ccallback.py',
   '_disjoint_set.py',
